requirements.txt
tensorflow
numpy
opencv-python
lxml
tqdm
flask
seaborn
pillow
